swap events for commands, optimize bindings

+++ GUI/ViewModels/BreedsViewModel.cs
@@ -1,4 +1,6 @@
 ﻿using System.Collections.Generic;
+using System.Net.Sockets;
+using System.Windows.Input;
 using GUI.Core;
 using Persistence.Models;
 using Persistence.Repositories;
@@ -10,7 +12,11 @@ public class BreedsViewModel : ViewModel
     private readonly BreedRepository _breedRepository;
     private IEnumerable<Breed> _breeds = new List<Breed>();
     private Breed? _selectedBreed;
-    private string _enteredName;
+    private string _name;
+
+    public ICommand AddBreedCommand { get; set; }
+    public ICommand UpdateBreedCommand { get; set; }
+    public ICommand DeleteBreedCommand { get; set; }
 
     public IEnumerable<Breed> Breeds
     {
@@ -32,9 +38,24 @@ public class BreedsViewModel : ViewModel
         }
     }
 
+    public string Name
+    {
+        get => _name;
+        set
+        {
+            _name = value;
+            OnPropertyChanged(nameof(Name));
+        }
+    }
+
     public BreedsViewModel(BreedRepository breedRepository)
     {
         _breedRepository = breedRepository;
+
+        AddBreedCommand = new RelayCommand(AddBreed, o => true);
+        UpdateBreedCommand = new RelayCommand(UpdateBreed, o => true);
+        DeleteBreedCommand = new RelayCommand(DeleteBreed, o => true);
+
         Breeds = _breedRepository.GetAll();
     }
 
@@ -43,33 +64,34 @@ public class BreedsViewModel : ViewModel
         SelectedBreed = breed;
     }
 
-    public void AddBreed()
+    public void AddBreed(object? unused)
     {
-        if (_enteredName.Length > 0)
+        if (Name?.Length > 0)
         {
-            Breed newBreed = new() { Name = _enteredName };
+            Breed newBreed = new() { Name = Name };
             _breedRepository.Add(newBreed);
             Breeds = _breedRepository.GetAll();
         }
     }
 
-    public void ChangeNameText(string name)
+    public void UpdateBreed(object? parameter)
     {
-        _enteredName = name;
+        if (parameter is Breed breed)
+        {
+            _breedRepository.Update(breed);
+            Breeds = _breedRepository.GetAll();
+            SelectedBreed = null;
+            SelectedBreed = _breedRepository.Get(breed.Id);
+        }
     }
 
-    public void UpdateBreed(Breed breed)
+    public void DeleteBreed(object? parameter)
     {
-        _breedRepository.Update(breed);
-        Breeds = _breedRepository.GetAll();
-        SelectedBreed = null;
-        SelectedBreed = _breedRepository.Get(breed.Id);
-    }
-
-    public void DeleteBreed(Breed breed)
-    {
-        _breedRepository.Delete(breed.Id);
-        Breeds = _breedRepository.GetAll();
-        SelectedBreed = null;
+        if (parameter is Breed breed)
+        {
+            _breedRepository.Delete(breed.Id);
+            Breeds = _breedRepository.GetAll();
+            SelectedBreed = null;
+        }
     }
 }
